add bare layout, configure new index page
