+++ layouts/bare.vue
@@ -1,0 +1,13 @@
+<template>
+  <div>
+    <slot />
+  </div>
+</template>
+
+<script lang="ts" setup>
+
+</script>
+
+<style>
+
+</style>
